fileVolume: Fix creation of tiny volumes
When creating preallocated volume with size <= 4096, we try to allocate
0 bytes:

    /usr/libexec/vdsm/fallocate --offset 4096 0 /path...

Which fail in fallocate() with EINVAL as expected, and then we raise

    VolumesZeroingError: Cannot zero out volume: /path...

It looks like engine does not handle this error correctly, and the disk
remain locked forever.

Creating a volume of 4k bytes or less is not a real use case, but there
is no technical reason why it should fail. Fixed by skipping the
allocation if we don't have anything to allocate.

Fixes: 84ef2dd05fe1 (fileSD: Do not use qemu-img preallocation=falloc)
Reported-by: Janos Bonic <jpasztor@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>