thinp: Fix handling of ENOSPC
Previously we called monitor_volumes(), but this does nothing unless a
drive block threshold is UNSET or EXCEEDED. If the guest is writing
fast, it may exceed the block threshold before we set it, and this case
we will never get a block threshold event from libvirt.

This can lead to paused VM with full disk that will never be extended.
Since the disk is full, resuming the VM fails immediately with ENOSPC.
But since block threshold was already exceeded, qemu does not submit a
block threshold event. The only way to recover is to shutdown the VM.

Fixed by unifying the way we handle ENOSPC and block threshold event; We
mark a drive block threshold as EXCEEDED in both cases. The periodic
monitor will pick up the drive in the next monitoring cycle.

Since on_enospc() cannot handle the case of ENOSPC without a drive, we
call it only when know the drive. Fixing monitor_volumes() to handle
this case is possible but require a big change, and I don't think this
case actually happens.

This change have nice side effects:

- Avoid the issue of blocking the libvirt event loop when calling
  monitor_volumes() and storage APIs block.

- Avoid noise in the logs when libvirt issue multiple ENOSPC events for
  the same drive. We log changes in drive block threshold only once.

This change introduces up to 2 seconds delay when resuming a paused VM.
However, since the VM already paused, the delay is not critical. This
delay will be eliminated when we eliminate the delay when receiving
block threshold event.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>