thinp: Fix forcing block threshold to exceeded
When finding that a drive need extension we force drive threshold state
to exceeded to make sure we continue to monitor this drive until it is
extended. This is good, but the way this was implemented is wrong,
clearing the exceeded time. Use Drive.on_block_threshold() which does
the right thing now.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>