thinp: Fix extending of exceeded volume
When libvirt report that the threshold was exceeded:

    2022-04-09 23:30:01,179+0300 INFO  (libvirt/events) [virt.vm]
    (vmId='724d41da-0b01-4d58-8a73-113906da1565') Block threshold
    3221225472 exceeded by 196608 for drive 'sdb[1]'

querying block status may show that allocation is smaller than the
threshold:

    2022-04-09 23:30:01,186+0300 DEBUG (periodic/1) [virt.vm]
    (vmId='724d41da-0b01-4d58-8a73-113906da1565') Extension info for
    drive sdb volume e1b76177-d925-497d-93ff-05ccb980f32a:
    BlockInfo(index=1, name='sdb', allocation=3213557760,
    capacity=53687091200, physical=5368709120, threshold=0)

This happens when trying to extend a disk immediately after receiving a
block threshold event.

Change the check to always extend drives when threshold state is
EXCEEDED without checking the amount of free space.

Since we always extend, we don't need the workaround for libvirt bug
when allocation is not reported.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>