tests: Improve optimal size tests for leaf volume
The first test checked the case when optimal size is actual_size +
chunk_size. But because the test use 1 GiB virtual size, the result may
be clipped by the maximum volume size. Change the test to use 2 GiB.

The second test test the case when size is limited by maximum size, by
writing 200 MiB to the volume. The test was correct, but since it
writes lot of data it was defined as slow tests, so it does not run in
the CI. Change the test to use 512 MiB volume, so we don't need to write
anything to test that the size is limited by the maximum size, and we
can run the tests everywhere.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>