blockVolume: Remove minimum internal volume size
When computing the size of internal volume, we rounded up to 1 GiB. I'm
not sure why we did this, since this does not help anything. The only
limit we have is lvm extent size. Adding extra limits only complicate
the code and make it harder to understand the system behavior.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>