blockVolume: Extract optimal_cow_size()
Before merging into base volume, we need to compute the optimal size of
the volume by measuring the volume, and adding free space. The
calculation is done today in BlockVolume.optimal_size(), but this uses
the actual size of the qcow2 image.

Extract BlockVolume.optimal_cow_size() accepting the required size, so
we can use the same calculation for the base volume before merge.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>