tests: Fix block volume tests fake chain
The tests created an invalid chain when the base volume is a leaf
volume. This error was hidden since the code does optimize for internal
merge, and always adds one chunk to the base.

Change make_env() so the caller can specify if the top volume is a leaf.
In this case the base volume will become the leaf after the merge.

Change the tests to create the Volume and Expected named tuples using
key=value style. This make the test setup much easier to follow.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>