merge: Fix over allocation when preparing
When preparing a merge, we calculate the new size of the base volume in
a dumb way:

    new_size = base_size + top_size + chunk_size

In the best case when merging a new empty snapshot, this allocates 2
extra chunks (5 GiB). In the worst case, when top is big, but all the
clusters in the top volume exist in the base volume, this can allocate
huge amount which is not needed.

The extra allocation is fixed after the merge, when we shrink the volume
to optimal size.

The extra allocation during the merge can cause:
- Failure to merge if the required allocation is more then the free
  space in the storage domain.
- VM can pause without a way to resume them if there is no space to
  extend thin disks
- Other storage operations may fail because there is not enough free
  space in the storage domain.

Fix the over allocation by measuring the required size for the merge,
and calculating the optimal size for the volume based on the required
size.

Since we calculate the optimal size before the merge, we can skip the
shrinking to optimal size at the end of the merge if the volume size
after the merge is optimal. This makes the merge more robust and little
faster.

Since current merge tests use fake block volumes, we must mock
qemuimg.measure() for testing leaf volumes. This should be improved
later by using a real block storage domain.

Fixes: #134
Signed-off-by: Nir Soffer <nsoffer@redhat.com>