blockVolume: Calculate optimal size as leaf
When finalizing a merge we calculate the optimal size for the base
volume in the wrong way, since optimal_size() is correct only after the
top volume was deleted, and the base volume becomes a leaf.

Add as_leaf argument to optimal_size(). When it is True, we calculate
the optimal size as if the volume is a leaf. Finalize merge will use
this to calculate correct size when finalizing an active merge.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>