merge: Fix reducing after merge
When reducing volume after an active merge, we reduced the base volume
too much, leaving no free space. When starting a VM after the merge, the
VM is very likely to pause with ENOSPC when the guest is writing to the
merged disk.

Fix by calculating the optimal size correctly, using the new as_leaf
argument. With this fix, the leaf volume after a merge always has one
chunk of free space (2.5g), and it behave like a new volume or snapshot.

Fixes: #179
Signed-off-by: Nir Soffer <nsoffer@redhat.com>