mailbox: Use same logger
The mailbox classes use logger per class. This is unneeded since we log
the thread names, and we move most code to module logger. To keep this
change minimal and easy to backport, I only change the logger names so
the class loggers use the same logger name "storage.mailbox".

We will switch to module logger later for oVirt 4.5.1.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>