mailbox: Silence common logs
The is no need to log the command sending mail to the SPM in INFO level,
and there is no need to log the command parameters since they are logged
by the mailbox command runner anyway.

The mailbox event logs are nice for debugging, but not needed in INFO
level.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>