thinp: Lock drive during monitoring
We want to extend drives as soon as possible when receiving events from
libvirt. However we may receive multiple events at the same time, or the
periodic monitor may wake up at the same time we want to handle an
event. Running the monitor concurrently can lead to double extend
messages and confusing logs.

Handle multiple calls by locking the drive during monitoring.  If
another thread try to monitor a drive while the drive monitor lock is
held, it waits up 0.5 seconds for the lock and skip the check on
timeout.

A new `thinp:monitor_timeout` option was added to control the timeout.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>