thinp: Fix race after a volume is refreshed
When a volume was refreshed, it was still marked as exceeded. If the
periodic montior wakes up at this point, it can try to extend the drive
again. This is unlikely event, but if it happens there is no way to fix
it without shutting down the VM.

Avoid the race by locking the drive monitor lock before the refresh, and
releasing the lock after we set a new block threshold. If the monitor
tries to access the drive at this point, it will wait until the
completion callback is completed or skip the check on timeout.

If we cannot acquire the drive monitor lock in 60.0 seconds to we abort
the refresh to avoid blocking the mailbox thread pool. This may happen
if the drive monitor lock is held by another monitor thread block on
inaccessible storage.

A new `thinp:refresh_timeout` option was added to control the timeout.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>