Consolidate semantically equal endpoint params
In 'copy_data', Endpoint attributes 'is_destination'
and 'writable' are semantically the same as both are
true only for the destination.

However, this behaviour could lead to a potential
programming bug. The attributes are not checked in the
body of the constructor, and a misusage could create a
non-writable destination endpoint, which is not correct.

To avoid this, remove 'writable' as a parameter of the
constructor and as a class member, and keep only 'is_destination',
deducted from the 'dest' parameter. This way we enforce
the correct behavior.

Fixes: #107
Signed-off-by: Albert Esteve <aesteve@redhat.com>