documentation: fixing the broken "git clone" command
After January's moving to GitHub, the clone VDSM fork command
was updated from "gerrit" to "github", but mistakenly the
"git clone" command itself was omitted.
Fixing to have the full command.
Also changed to clone using HTTPS protocol in order to be consistent
with other projects' clone commands in the same document.

Signed-off-by: Pavel Bar <pbar@redhat.com>