documentation:  using  "shell" format to highlight snippets with commands
Using a built-in Markdown "shell" format to beautifully highlight
snippets with commands to be executed at CLI.
This format is presented really neat in PyCharm IDE, which even knows to display
the preceding "$" in a different color and execute the command(s) in the
built-in IDE terminal (could be useful if we executed PyCharm on the VDSM host).
Also added a preceding "$ " sign to resemble a real CLI look-and-feel.

Signed-off-by: Pavel Bar <pbar@redhat.com>