thinp: Minimize duplicate extend requests
To allow immediate extend when receiving a block threshold or enospc
events, we need to handle the case when the monitor wakes up soon after
we extended the drive, find that the drive is exceeded, and try to
extend it again. This creates duplicate extend requests that should be
harmless, but they make it hard to analyze extend stats.

To avoid this issue we keep the time of the last extend in the drive.
When the monitor handles an exceeded drive, it skips this drive if a
previous extend was started recently, and the extend timeout did not
expired.

We need to handle also another case, when a drive was extended recently,
but setting the block threshold failed, and it should be extended again.
If the guest write quickly before the extend timeout expired, we want to
extend the drive immediately.

To allow immediate extend, add an "urgent" argument to
_handle_exceeded(). It is set when we find that the drive needs to be
extended in _handle_unset(). It will be also used when receiving a block
threshold or enospc events.

New `thinp:extend_timeout` option added to control the timeout.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>