thinp: Support immediate extend on events
If the monitor is configured with an executor dispatch function, it
dispatches a call to extend the drive when getting a block threshold or
enospc events. The drive will be extended on the first available
executor worker.

Using a dispatch function keeps the thinp module decoupled from the
periodic module, makes testing easy, and prepares for removing the
periodic monitor and use thinp executor.

The periodic monitor checks if VM is ready for commands before
monitoring. Testing extending of 4 drives at the same time shows that
this check returns False randomly for no reason. Since this cause a
delay before extending the drive, we don't do this check when receiving
an event.

We also don't check migration status or use it when handling libvirt
calls failures, since it should be needed for volume monitoring,
disabled on migration destination VM.

When extending a drive we hold the drive monitor lock, to avoid
conflicts with the periodic monitor and extend completion threads.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>