thinp: Fix handling of drive extended to maximum size
The recent change to log a warning when exceeded drive cannot be
extended since it was already extended to the maximum size revealed that
we try to monitor and extend such drive every 2 seconds.

Example failure flow with 20g drive:

1. Vdsm extends drive to 22g, and set threshold to 20g
2. Guest writes to offset 20g
3. Vdsm get block threshold event and mark drive as exceeded
4. Every 2 seconds, vdsm try to extend the drive and log a warning:

    WARN  (periodic/0) [virt.vm] (vmId='d6eb4739-ccd3-4652-b6fc-7f4fd4a972ad')
    Drive already extended to maximum size ...

Why we extend the drive to 22g if drive capacity is 20g?
When using qcow2 format, writing 20g of guest data requires additional
space for qcow2 metadata. Vdsm allows up to 1.1 * capacity (22g).

This is not a new issue - it existed since we started to use block
threshold few years ago. But since we did not log anything, the bug was
hidden, consuming CPU cycles and increasing the chance for filling up
the executor queue and delaying other tasks.

To fix this issue, introduce a new threshold state: `DISABLED`. In this
state the drive is not picked up for monitoring. If the drive is
resized, we change the threshold state back to `UNSET` so the drive will
be monitored again in the next monitor cycle.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>