lvm: improve error handling for single vg getter
Currently, if we try to reload a non existing VG that is not
in the cache, we will not get any error, it will raise an
exception without any context of the actual reason of the fail.

Furthermore, if the LVM command fails, it might go unnoticed,
or trigger a single uninformative warning without the reason
why it might have failed.

Make _reload_single_vg to raise an exception if the LVM command
failed and there is no output or the VG is stale (not updated).
Otherwise it returns the single VG that has been updated.
LVMCache.getVg() will propagate the exception.

Signed-off-by: Albert Esteve <aesteve@redhat.com>