lvm: simplify listPVnames
listPVnames first tries to get the VG directly from
the cache in lvminfo, without using the getter, thus
ignoring the case of a stale or unreadable VG.

Then tries to catch those cases with a except block
catching an AttributeError or a KeyError, and recover
from the issue by calling getVG.

However, we should simply use getVG from the get-go,
which already handles missing, stale or unreadable VG,
and tries to reload it from LVM. If it fails it will
raise a VolumeGroupDoesNotExist.

Also, VGBlockSizes methods were checking missing PVs
in a VG after calling listPVnames, and raising a
VolumeGroupDoesNotExist if empty.

But there is no such thing as a VG without PVs (pv_name),
We either found a VG with a valid pv_name, or listPVnames
rises an exception, and therefore we can remove that part.

Signed-off-by: Albert Esteve <aesteve@redhat.com>