lvm_test: update tests to run on vdsm hosts
Some commands run and try to change configuration for the
lvm command with the --config option, but this is wrong
for hosts configured for vdsm, since they use devicesfile
by default. Tests should be adapted and use --devices.

This happens mostly in fixtures simulating stale devices,
but other tests are also affected. Fix all occurrences and
make them use `--devices` instead of `--config`.

Fixes: #163
Signed-off-by: Albert Esteve <aesteve@redhat.com>