iscsi: Keep existing session on "session exists"
If logging in to a node fails with "session exists" (error 15):

    iscsiadm: initiator reported error (15 - session exists)

We use to remove the node, which disconnects the node and remove it.
This is not new behavior, but it seems that in 4.4 this cleanup was not
effective in the case of logging in to the same connection more than
once, and now it reliably disconnect the first node and leave the host
without any nodes, which makes it non operational.

Change iscsiadm to raise new IscsiSessionExists error, and keep the
existing session when handling this error.

With this change, if you try to connect to the same target more than
once, the host should end with one connected target.

Bug-Url: https://bugzilla.redhat.com/2083271
Signed-off-by: Nir Soffer <nsoffer@redhat.com>