net: Fix the tuple assign
The tuple returned from split_switch_type
was not properly assigned to variables.
This caused the if statement to be always true,
it did not cause any issues, but the network capabilities
were also getting OvS info which wasn't necessary.

Assign it properly so only the first dictionary is used.

Signed-off-by: Ales Musil <amusil@redhat.com>