merge: check can_reduce in a cold merge
When finalizing a merge, only the format (cow)
and the optimal_size are considered to discern when to
reduce a volume. However, calculating the optimal size can
be avoided in the case of a preallocated block volume, as it
cannot be reduced.

Use the volume can_reduce check before obtaining the optimal
size, and before reducing the volume, so that we consider the
type of a volume before reducing it.

Bug-Url: https://bugzilla.redhat.com/2081493
Signed-off-by: Albert Esteve <aesteve@redhat.com>