volume: add 'can_reduce' method
Currently there is not check performed from within
the volume class before a reduce is attempted. This may
result in reduce calls with wrong volume type (i.e., preallocated).

In preparation, provide a 'can_reduce' method in
the volume base class, so that it could be called from the
appropiate place before a volume is reduced.

Signed-off-by: Albert Esteve <aesteve@redhat.com>