lvm: split reloadvgs for single and multi VGs
LVMCache._reloadvgs is used the same way both when
reloading a single VG, and when reloading a set of VGs
or all VGs in the cache.

This makes it difficult to have tailored handling (e.g., error
handling) of reloading VGs for those cases. Also forces
the method to be overloaded, supporting different types
from the parameter and increasing complexity.

Specialize _reloadvgs for single VG invocation to
allow different behaviour and avoid function overloading.

Signed-off-by: Albert Esteve <aesteve@redhat.com>