blockSD: avoid argument shadowing in validatePVs
BlockStorageDomainManifest._validatePVsPartOfVG receives
a 'pv' argument that is shadowed in a dict comprehension
in the first line of the method. It might be safe as it
is contained in the comprehension only, but shadowing
should be avoided anyway.

Also, 'pv' is wrong in the comprehension, since listPVNames
will return a PV path (not a PV namedtuple). So changing
'pv' for 'path' in the comprehension will both fix the
argument shadowing and use proper informative naming.

Signed-off-by: Albert Esteve <aesteve@redhat.com>