virt: Prolong migration listener timeout by disk preparation
When a migrating VM is being prepared on the destination host, some
disks require modifying and reloading udev rules on the host.  This
can take a significant amount of time and if the VM has a lot of such
disks then a migration preparation timeout may occur.

The right solution would be to reload the modified rules only once for
all the disks.  But this would require some more work to implement.
At the moment, let’s make a workaround prolonging the migration
preparation timeout per each such a disk, i.e. a disk with
/dev/mapper/… source paths.  But not beyond a newly configurable
max_migration_listener_timeout.

We add 2 seconds per drive by default.  This is an arbitrarily
selected value that would work well in an environment where the
problem was observed.  If this value doesn’t work well enough in some
environment, it can be overridden in another newly introduced
configuration value migration_listener_prepare_disk_timeout.

Bug-Url: https://bugzilla.redhat.com/2055905