livemerge: Fix over extend
When extending the base volume before merge, we use a dumb calculation
extending the base volume by top_size + chunk_size. This allocate way
too much space which is typically not needed. For active layer merge,
there is no way to reduce the volume after the merge without shutting
down the VM. The result is growing the active volume on every merge,
until it consumes the maximum size.

Fix the issue by measuring the sub-chain from top to base before the
extend. This give the exact size needed to commit the top volume into
the base volume, including the size required for the bitmaps that may be
in the top and base volume.

In the case of active layer merge, this measurement is a heuristic,
since the guest can write data during the measurement, or later during
the merge. We add one chunk of free space to minimize the chance of
pausing a VM during merge. The only way to prevent pausing during merge
is to monitor base volume block threshold during the merge.  This was
not possible in the past, but can be done with current libvirt, but vdsm
thin provisioning code is not ready for this yet.

For internal merge, measuring is exact, and there is no need to leave
free space in the base volume since the top volume is read only.

Fixes: #188
Related-to: https://bugzilla.redhat.com/1993235
Signed-off-by: Nir Soffer <nsoffer@redhat.com>