lvm: support only reloading all LVs
LVMCache._reloadlvs was coded to support reloading
a single LV, a subset of all LVs in the VG, or all
of them. However, it is only used for reloading all,
since the cost is roughly the same.

Support only the case that is used, where all LVs
are to be reloaded avoids the unnecessary complexity,
and overloading the function signature with different
parameter types.

Signed-off-by: Albert Esteve <aesteve@redhat.com>