lvm: reloadlv raise LVDoesNotExistError
Move the responsibility of raising LogicalVolumeDoesNotExistError
to LVMCache class to make it consistent with VG error
handling.

Avoid raising if no Volumes are found when getting all
the LVs of a VG, as this is wrong. Only raise when looking
for a specific LV and not finding it. An empty VG is a valid
case (e.g., a recently created VG), and is the caller who
should decide whether finding no LVs is valid or not.

Fixes: #211
Signed-off-by: Albert Esteve <aesteve@redhat.com>