lvm: reloadpv raise InaccessiblePhyDev
Currently the PV getter raises InaccessiblePhyDev
if None is detectedi, which is the default return value
of the PVs dictionary when the key does not exist.

Move the raise to the reload method, so that we raise
as soon as the missing PV is detected and we leave
the responsibility of raising always to the LVMCache,
consistent with how VG errors are handled.

Signed-off-by: Albert Esteve <aesteve@redhat.com>