vm: Simplify extend_volume()
Previously we accepted the current and maximum size of the volume, and
computed the next volume size, including one chunk of free space, and
ensuring that the size is limited by the volume capacity and qcow2
overhead.

This works for auto extending thin volume and for pre extend during
replication, but it does not work correctly for internal live merge,
when we don't want to add one chunk of free space. It also ugly that in
active layer merge we have to lie and call with the maximum required
size as the "current" size.

Change extend_volume() to accept the new size, moving the responsibility
to the caller. The 3 callers were update to compute the size using
drive.getNextVolumeSize().

Signed-off-by: Nir Soffer <nsoffer@redhat.com>