lvm: split LV getter for single or all LVs
Currently lvm.getLV is multipurposed for single
and multi LV lookup, which is also replicated
in LVMCache.getLv method. This makes it difficult
to handle these two cases properly and raise
if an error occurs looking up a single LV.

Split lvm.getLV function into single LV getter,
that requires an LV name, and a getter for all
the LVs in a VG, which will only require the VG
name.

Replicate the split in LVMCache LV getters.

Signed-off-by: Albert Esteve <aesteve@redhat.com>