tox.ini: allow newer black version to match click
Latest click versions are not compatible with black
version used by tox (21.6b0). To fix it we either
have to move to a more recent version of black, which
currently breaks the linter in some tests, or we downgrade
the version of click to 8.0.2.

Update black version to a more recent one (22.3.0) and fix
linting issues that come with the change.

Signed-off-by: Albert Esteve <aesteve@redhat.com>