virt: use proper CPU pinning when resuming hibernated VM
When hibernating a VM we store also the CPU pinning. This pinning needs
to be invalidated when resuming the VM because in most cases (except
when manual pinning is involved) the pinning is no longer usable either
because shared pool needs to be adapted or dedicated CPUs have changed.

New pinning passed from engine needs to be used instead.

Domain XML that is used to restore the VM is loaded from libvirt's
hibernation image which replaces XML passed from API call and that goes
through init. This should not lead to any change in behavior. While we
used the passed XML to create the domain it has been always immediately
replaced by libvirt in the resume call. That means if we are doing any
changes to the XML during early init such changes were always lost and
will still remain lost. It may be tempting to pass domain XML from
engine or from our pickled data to `resumeFlags()`. Neither is possible
because our domain XML is not migration XML. Doing so would end in error
because libvirt would fail to match the guest specific portion with the
stored domain XML.

Bug-Url: https://bugzilla.redhat.com/2083302
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>