storage.exception: add _HoldingLVMCommandError
Add _HoldingLVMCommandError to act as a parent class for all
StorageExceptions that occur as a consecuence of an LVM
command error. This class is not meant to be directly
instantiated.

The exception holds an Optional LVMCommandError, validates
the parameter in the constructor, and overwrites the `value`
property for logging information consistently.

Signed-off-by: Albert Esteve <aesteve@redhat.com>