check: Fix fd leak when child process fails
When using Popen.communicate() the process pipes are closed when the
child closes its end of the pipe. But in the check module we run
multiple processes in the same thread, and we read from the pipe using
asynchronous I/O. In this case we never close the child process pipe.
This should not be an issue, since the pipe should be closed
automatically when we remove the reference to the process, and we never
notice any file descriptor leak in this code.

However we see clear leak in the case when a storage domain becomes
inaccessible. Strangely, only failing checks from inaccessible storage
domains leak, and we see very clear leak of one file descriptor every 10
seconds:

    for i in $(seq 3600); do
        echo "$(date --rfc-3339=seconds), $(readlink /proc/$(pidof -x vdsmd)/fd/* | wc -l)"
        sleep 10
    done
    ...
    2022-06-07 22:52:14+03:00, 374
    2022-06-07 22:52:24+03:00, 375
    2022-06-07 22:52:34+03:00, 376
    2022-06-07 22:52:44+03:00, 377
    2022-06-07 22:52:54+03:00, 378
    2022-06-07 22:53:04+03:00, 379
    2022-06-07 22:53:14+03:00, 379
    2022-06-07 22:53:24+03:00, 380
    2022-06-07 22:53:34+03:00, 381
    2022-06-07 22:53:44+03:00, 384
    2022-06-07 22:53:54+03:00, 383
    2022-06-07 22:54:04+03:00, 384
    2022-06-07 22:54:14+03:00, 385
    2022-06-07 22:54:24+03:00, 386
    2022-06-07 22:54:34+03:00, 387
    2022-06-07 22:54:44+03:00, 388
    2022-06-07 22:54:54+03:00, 388
    2022-06-07 22:55:04+03:00, 389

I cannot explain why this happen only when the process fail and we read
an error message from stderr.

Fix by explicitly closing the child process stderr pipe when we finish
reading from it.

Bug-Url: https://bugzilla.redhat.com/2075795
Signed-off-by: Nir Soffer <nsoffer@redhat.com>