lvm: add check for vgs autoactivation field
VGs in LVM are created with disabled autoactivation in Vdsm to avoid
unwanted activation of VG/LVs when the autoactivation option (-aay) is
used. Autoactivation is used, e.g., during system boot.

However, this configuration can be overridden, either in the lvm.conf or
by manually invoking a command that overwrites the VG settings.

Add a check check for the autoactivation field of the VGs. If any VG
has the attribute enabled, log a warning and deactivate it.

Signed-off-by: Albert Esteve <aesteve@redhat.com>