sp: check for VGs autoactivation
Changing VGs can be done only the SPM host. Other hosts
should not check or warn about the VG since they cannot
fix this issue.

Make the SPM use the LVM check that fixes the issue when enabled.
The check is used when starting the SPM and when activating a
storage domain.

Fixes: #237
Signed-off-by: Albert Esteve <aesteve@redhat.com>