lvmlocal.conf: Disable multipath_wwids_file usage
In RHEL 8.6, LVM added detection of multipath components using the
multipath wwids file (/etc/multipath/wwids). Unfortunately this
incorrectly detect blacklist multipath devices. This causes
vgimportdevices to skip the devices when adding to the devices file, and
the result is host failing to boot.

Since we always use lvm devices or filter, this feature is not helpful
for our use case, and can be safely disabled.

Bug-Url: https://bugzilla.redhat.com/2090169
Related-to: https://bugzilla.redhat.com/2076262

Signed-off-by: Nir Soffer <nsoffer@redhat.com>