sp: directly import lvm module
In sp module, storage.lvm was used through the BlockSD import
namespace, which itself includes the lvm module. This is ok-ish
as long as it is only used once in the module, but in general
not recommended, since it obscures which lvm module are
we using, and creates a weird namespace pattern if we want
to use lvm in more places.

Better to import the lvm module directly in sp.

Signed-off-by: Albert Esteve <aesteve@redhat.com>