virt: fix graphics configuration when resuming from hibernation
Somehow libvirt drops `passwdValidUntil` attribute from domain XML when
when we pass XML that has no `passwd` attribute to restoreFlags(). Same
happens when restoring from snapshot with memory and we can use the same
method to mitigate the problem.

An alternative would be to pass `VIR_DOMAIN_SAVE_IMAGE_XML_SECURE` flag
to `saveImageGetXMLDesc()` call. That way libvirt would not hide the
`passwd` from us. But since our passwords to graphics console are
temporary this does not add any value. Also we would leak the console
password to the logs which is not favorable even though the password is
likely to be invalid at the time of resume.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>