hsm: connectStorageServer receive device list
When connecting a device to a host, add an option for the
caller to specify the list of devices to wait for in
StoragePool.connectStorageServer.

By default, Vdsm uses udevadm settle and 5 seconds timeout
to make sure devices are settled before continuing after
setting up an iscsi device, but that is still not robust
enough and can break due to a race condition.

By telling Vdsm which devices must exist on the host
when connecting to a server for making a direct LUN
available, we can make the API more robust.
Vdsm will check for the device wwid to be available
in the /dev/mapper folder, and then wait for background
events to be finished before proceeding.

Bug-Url: https://bugzilla.redhat.com/2074029
Signed-off-by: Albert Esteve <aesteve@redhat.com>