lvm: move LVMRunner to new lvmcmd module
Create a new lvmcmd module to hold all lvm command
logic. Start with moving LVMRunner class to lvmcmd
so that running the command is moved out of the lvm
module.

Final goal is leave lvm module only for
handling devices and the LVMCache, and move all the
running and formatting lvm command to lvmcmd.

Signed-off-by: Albert Esteve <aesteve@redhat.com>