lvm: move LVM formatting methods to lvmcmd
Move LVM formatting methods from LVMCache to lvmcmd
module. This way the formatting in untangled from the
cache class, and organized together in its own wrapper
module. LVMCache does not need to own the LVMRunner
anymore.

Create new test fixture for lvm_test so that the
fake_runner can be updated in one single place,
and use it whenever a fake_runner was created.

Signed-off-by: Albert Esteve <aesteve@redhat.com>