mpathconf: use readline for config read
In the multipath configuration revision check,
the whole file is read with readlines(), but
only the first two lines are used. If the file
is empty, the list will be empty causing a posible
KeyError that needs to be handled.

All this complexity can be saved by using readline(),
to read only the two lines needed. It already gives
an empty string as a default value when the end of
the file is reached, so we don't need to handle any
exception, simplifying the code and improving
readability.

Finally, the lines read used rstrip for eliminating
trailing newlines from the string. This can now be
skipped since we use 'startswith' to check line,
which exits as soon as the check is either
positive or negative, and therefore it will never
reach the newline character.

Signed-off-by: Albert Esteve <aesteve@redhat.com>